refactor(domain): 도메인 패키지에서 다시 model과 repository 패키지로 나눔
