+++ src/main/java/kitchenpos/products/tobe/domain/model/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.domain.model;
 
 import java.math.BigDecimal;
 import java.util.UUID;
