+++ src/main/java/kitchenpos/products/tobe/infra/JpaProductRepository.java
@@ -1,8 +1,8 @@
 package kitchenpos.products.tobe.infra;
 
 import java.util.UUID;
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
+import kitchenpos.products.tobe.domain.model.Product;
+import kitchenpos.products.tobe.domain.repository.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaProductRepository extends ProductRepository, JpaRepository<Product, UUID> {
