+++ src/test/java/kitchenpos/products/tobe/domain/DisplayedNameTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import kitchenpos.products.application.FakePurgomalumClient;
+import kitchenpos.products.tobe.domain.model.DisplayedName;
 import kitchenpos.products.tobe.exception.WrongDisplayedNameExeption;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
